test: one file per router — test_google.py absorbs test_editor_apis.py
Gmail, Drive and the Workspace editor reads are one router (app/routers/google.py) and one error
envelope, so they are one test file. test_editor_apis.py — named for the campaign that added it,
not for what it tests — is gone; its 44 tests were 34 Drive/editors and 9 cross-cutting.

_tok moves to _helpers as tok() (notion uses it too). Dependency-checked both ways before moving:
no kept test references a moved name and vice versa.

test_endpoints.py 3266 -> 2270 lines. Same 860 passing.
