+++ tests/test_endpoints.py
@@ -23,36 +23,6 @@ from app.config import Settings
 from tests._helpers import build_corpus, client_for, db_count
 
 
-# --- crawlers (small page sizes to exercise pagination) -------------------------
-
-def crawl_gmail(client, headers, user="me"):
-    ids, token = [], None
-    while True:
-        p = {"maxResults": 7}
-        if token:
-            p["pageToken"] = token
-        j = client.get(f"/gmail/v1/users/{user}/messages", headers=headers, params=p).json()
-        ids += [m["id"] for m in j.get("messages", [])]
-        token = j.get("nextPageToken")
-        if not token:
-            break
-    return ids
-
-
-def crawl_drive(client, headers):
-    ids, token = [], None
-    while True:
-        p = {"pageSize": 7}
-        if token:
-            p["pageToken"] = token
-        j = client.get("/drive/v3/files", headers=headers, params=p).json()
-        ids += [f["id"] for f in j.get("files", [])]
-        token = j.get("nextPageToken")
-        if not token:
-            break
-    return ids
-
-
 def crawl_github_repo(client, headers, org, repo):
     out, page = [], 1
     while True:
@@ -145,19 +115,6 @@ def crawl_slack(client, headers):
             if not ccur:
                 break
     return total
-
-
-# --- admin full-crawl completeness ---------------------------------------------
-
-def test_admin_gmail_crawls_all(client, admin_h, ro_conn):
-    assert len(crawl_gmail(client, admin_h)) == db_count(ro_conn, "gmail")
-
-
-def test_admin_drive_crawls_all(client, admin_h, ro_conn):
-    # An unfiltered files.list includes folders on real Drive, and the mock synthesizes one per
-    # container — so a full crawl is every stored file plus every folder.
-    folders = ro_conn.execute("SELECT COUNT(*) FROM gdrive_folders").fetchone()[0]
-    assert len(crawl_drive(client, admin_h)) == db_count(ro_conn, "google_drive") + folders
 
 
 def test_admin_github_crawls_all(client, admin_h, ro_conn, org):
@@ -523,224 +480,6 @@ def test_hubspot_batch_read(client, admin_h):
                     json={"inputs": [{"id": i} for i in ids],
                           "properties": ["name"]}).json()
     assert {r["id"] for r in j["results"]} == set(ids)
-
-
-# --- content round-trips through each vendor's encoding -------------------------
-
-def _gmail_plain(payload):
-    """Extract the text/plain body data from a Gmail payload (top-level or a part)."""
-    if payload.get("body", {}).get("data"):
-        return payload["body"]["data"]
-    for part in payload.get("parts", []):
-        if part["mimeType"] == "text/plain":
-            return part["body"]["data"]
-    raise AssertionError("no text/plain part")
-
-
-# --- Gmail hex message ids (#39) --------------------------------------------------------------
-#
-# Gmail ids are 16 lowercase hex digits parsed as a signed 64-bit integer. The 400/404 boundary,
-# MEASURED against the live API:
-#
-#   id                        | real Gmail
-#   --------------------------|-----------------------------------------------
-#   0 / 1 / abc123 / DEADBEEF | 404 NOT_FOUND     (a valid shape, just unknown)
-#   7fffffffffffffff          | 404 NOT_FOUND     (2**63 - 1 is in range)
-#   8000000000000000          | 400 INVALID_ARGUMENT "Invalid id value"
-#   ffffffffffffffff          | 400               (>= 2**63)
-#   18c9a1b2c3d4e5f6a         | 400               (17 digits overflows)
-#   -1 / 1g / " 1"            | 400               (not hex)
-#
-# Threads share the id space: a single-message thread reports id == threadId.
-
-def _a_gmail_row(ro_conn):
-    return ro_conn.execute("SELECT * FROM gmail_messages LIMIT 1").fetchone()
-
-
-def test_gmail_messages_list_serves_hex_ids(client, admin_h):
-    """The ids a client receives must look like Gmail's, not like the corpus's dsids — that is the
-    whole point of #39. `dsid_…` is not hex, so real Gmail would call it an invalid id value."""
-    msgs = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                      params={"maxResults": 10}).json()["messages"]
-    assert msgs
-    for m in msgs:
-        for key in ("id", "threadId"):
-            assert len(m[key]) == 16, m
-            assert all(c in "0123456789abcdef" for c in m[key]), m
-            assert int(m[key], 16) < 2 ** 63, m
-        assert not m["id"].startswith("dsid_")
-
-
-def test_gmail_hex_id_resolves_to_the_same_document(client, admin_h, ro_conn):
-    """The hex id maps back to its dsid, so the body a client reads by hex is the stored body. A
-    one-way id would make every message unreadable."""
-    from app import synth
-    row = _a_gmail_row(ro_conn)
-    hexid = synth.gmail_message_id(row["doc_id"])
-    m = client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=admin_h,
-                   params={"format": "full"}).json()
-    assert m["id"] == hexid
-    assert base64.urlsafe_b64decode(_gmail_plain(m["payload"])).decode() == row["content"]
-
-
-def test_gmail_thread_id_matches_the_message_id_for_a_lone_message(client, admin_h, ro_conn):
-    """Threads share the message id space in real Gmail, so a message that is its own thread root
-    reports the same value twice — and `threads.get` resolves it."""
-    from app import synth
-    row = ro_conn.execute(
-        "SELECT * FROM gmail_messages WHERE COALESCE(thread_id, '') = '' LIMIT 1").fetchone()
-    if row is None:
-        row = ro_conn.execute(
-            "SELECT * FROM gmail_messages WHERE thread_id = doc_id LIMIT 1").fetchone()
-    assert row is not None, "SAMPLE should hold a message that is its own thread"
-    hexid = synth.gmail_message_id(row["doc_id"])
-    m = client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=admin_h).json()
-    assert m["id"] == m["threadId"] == hexid
-    t = client.get(f"/gmail/v1/users/me/threads/{hexid}", headers=admin_h)
-    assert t.status_code == 200 and t.json()["id"] == hexid
-
-
-def test_gmail_reply_reports_its_roots_thread_id(client, admin_h, ro_conn):
-    from app import synth
-    row = ro_conn.execute("SELECT * FROM gmail_messages WHERE COALESCE(thread_id,'') != '' "
-                          "AND thread_id != doc_id LIMIT 1").fetchone()
-    assert row is not None, "SAMPLE should hold a threaded reply"
-    m = client.get(f"/gmail/v1/users/me/messages/{synth.gmail_message_id(row['doc_id'])}",
-                   headers=admin_h).json()
-    assert m["threadId"] == synth.gmail_message_id(row["thread_id"])
-    assert m["id"] != m["threadId"]
-
-
-def test_gmail_attachment_resolves_under_a_hex_message_id(client, admin_h, ro_conn):
-    from app import synth
-    row = ro_conn.execute("SELECT * FROM gmail_messages WHERE COALESCE(attachments,'') NOT IN "
-                          "('', '[]') LIMIT 1").fetchone()
-    assert row is not None, "SAMPLE should hold a message with an attachment"
-    hexid = synth.gmail_message_id(row["doc_id"])
-    m = client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=admin_h,
-                   params={"format": "full"}).json()
-    att = next(p for p in m["payload"]["parts"] if p.get("filename"))
-    r = client.get(f"/gmail/v1/users/me/messages/{hexid}/attachments/"
-                   f"{att['body']['attachmentId']}", headers=admin_h)
-    assert r.status_code == 200 and r.json()["size"] > 0
-
-
-@pytest.mark.parametrize("mid", ["0", "1", "abc123", "DEADBEEF", "7fffffffffffffff",
-                                 "0000000000000001", "18c9a1b2c3d4e5f6"])
-def test_gmail_a_valid_but_unknown_id_is_not_found(client, admin_h, mid):
-    """A well-formed id the mailbox does not hold is 404, uppercase included — measured."""
-    for kind in ("messages", "threads"):
-        r = client.get(f"/gmail/v1/users/me/{kind}/{mid}", headers=admin_h)
-        assert r.status_code == 404, f"{kind}/{mid}: {r.status_code}"
-        assert r.json()["error"]["message"] == "Requested entity was not found."
-
-
-@pytest.mark.parametrize("mid", ["8000000000000000", "ffffffffffffffff", "18c9a1b2c3d4e5f6a",
-                                 "-1", "1g", "nosuchmessageid", "dsid_00908a2dda4b4d359194a09101"])
-def test_gmail_an_unparsable_id_is_an_invalid_argument(client, admin_h, mid):
-    """The gap #39 names: an id that is not a parsable in-range hex integer is 400
-    INVALID_ARGUMENT "Invalid id value", not 404. The last row is the mock's OWN former id format,
-    which is exactly why the served ids had to change first."""
-    for kind in ("messages", "threads"):
-        r = client.get(f"/gmail/v1/users/me/{kind}/{mid}", headers=admin_h)
-        assert r.status_code == 400, f"{kind}/{mid}: {r.status_code}"
-        e = r.json()["error"]
-        assert e["message"] == "Invalid id value"
-        assert e["status"] == "INVALID_ARGUMENT"
-        assert e["errors"][0]["reason"] == "invalidArgument"
-
-
-def test_gmail_hex_ids_still_enforce_the_acl(client, admin_h, tokens_yaml, ro_conn):
-    """Resolving through the index must not become a way around the ACL. The index is global — it
-    maps every hex id, visible or not — so the ACL read after it is the only thing standing between
-    a scoped caller and someone else's mail. The CFO's comp review is granted to cfo alone."""
-    from app import synth
-    row = ro_conn.execute("SELECT * FROM gmail_messages WHERE title LIKE 'Confidential comp%'"
-                          ).fetchone()
-    hexid = synth.gmail_message_id(row["doc_id"])
-    assert client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=admin_h).status_code == 200
-    cfo = {"Authorization": f"Bearer {_tok(tokens_yaml, 'cfo@acme.com')}"}
-    assert client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=cfo).status_code == 200
-    outsider = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    r = client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=outsider)
-    assert r.status_code == 404
-    assert r.json()["error"]["message"] == "Requested entity was not found."
-
-
-def test_gmail_body_roundtrip(client, admin_h, ro_conn):
-    from app import synth
-    doc = ro_conn.execute("SELECT * FROM gmail_messages LIMIT 1").fetchone()
-    m = client.get(f"/gmail/v1/users/me/messages/{synth.gmail_message_id(doc['doc_id'])}",
-                   headers=admin_h, params={"format": "full"}).json()
-    body = base64.urlsafe_b64decode(_gmail_plain(m["payload"])).decode()
-    assert body == doc["content"]
-    subj = next(h["value"] for h in m["payload"]["headers"] if h["name"] == "Subject")
-    assert subj == doc["title"]
-
-
-def test_gmail_messages_list_ordered_by_internaldate_desc(client, admin_h, ro_conn):
-    # Real Gmail returns messages.list newest-first by internalDate. Regression (#11): the mock
-    # listed by doc_id (hash order), so a capped "newest N" was effectively random by date.
-    listed = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                        params={"maxResults": 50}).json()["messages"]
-    got = [m["id"] for m in listed]
-    # the stable total order the endpoint must produce: created_ts DESC, doc_id ASC as tie-break
-    # the served ids are hex (#39), so the expectation is the hex of that stable order
-    from app import synth
-    expected = [synth.gmail_message_id(r["doc_id"]) for r in ro_conn.execute(
-        "SELECT doc_id FROM gmail_messages ORDER BY created_ts DESC, doc_id LIMIT 50").fetchall()]
-    assert got == expected
-    # ...and internalDate is monotonically non-increasing across the returned page
-    dates = [int(client.get(f"/gmail/v1/users/me/messages/{i}", headers=admin_h,
-                            params={"format": "minimal"}).json()["internalDate"]) for i in got]
-    assert dates == sorted(dates, reverse=True)
-
-
-def test_gmail_messages_list_pagination_stable_and_ordered(client, admin_h, ro_conn):
-    # Paging must be a stable partition of the same date-desc order — no dupes, no skips, and page 2
-    # continues strictly at/under page 1's tail. (Regression guard for the tie-break in ORDER BY.)
-    total = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                       params={"maxResults": 1}).json()["resultSizeEstimate"]
-    if total < 2:
-        pytest.skip("need >= 2 gmail messages to exercise paging")
-    p1 = client.get("/gmail/v1/users/me/messages", headers=admin_h, params={"maxResults": 1}).json()
-    p2 = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                    params={"maxResults": 1, "pageToken": p1["nextPageToken"]}).json()
-    a, b = p1["messages"][0]["id"], p2["messages"][0]["id"]
-    assert a != b                                                     # distinct rows, no repeat
-    both = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                      params={"maxResults": 2}).json()["messages"]
-    assert [m["id"] for m in both] == [a, b]                          # pages concatenate in order
-
-
-def test_gmail_attachment_size_matches_part_metadata(client, admin_h, ro_conn):
-    # Real Gmail's contract: a part's body.size equals the byte length attachments.get serves, so a
-    # client can stat an attachment from message metadata alone. Regression: the part reported the
-    # corpus-declared `size` (e.g. 2048) while attachments.get returned len(content) — a mismatch.
-    row = ro_conn.execute(
-        "SELECT doc_id FROM gmail_messages WHERE attachments IS NOT NULL "
-        "AND attachments != '[]' LIMIT 1").fetchone()
-    if row is None:
-        pytest.skip("no gmail message with an attachment in this subset")
-    from app import synth
-    hexid = synth.gmail_message_id(row["doc_id"])
-    m = client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=admin_h,
-                   params={"format": "full"}).json()
-    parts = [p for p in m["payload"]["parts"] if p.get("body", {}).get("attachmentId")]
-    assert parts, "message should expose at least one attachment part"
-    for p in parts:
-        got = client.get(
-            f"/gmail/v1/users/me/messages/{hexid}/attachments/{p['body']['attachmentId']}",
-            headers=admin_h).json()
-        assert got["size"] == p["body"]["size"]                       # the two agree
-        assert len(base64.urlsafe_b64decode(got["data"])) == p["body"]["size"]  # ...and match the bytes
-
-
-def test_drive_export_roundtrip(client, admin_h, ro_conn):
-    doc = ro_conn.execute("SELECT * FROM gdrive_files LIMIT 1").fetchone()
-    text = client.get(f"/drive/v3/files/{doc['doc_id']}/export", headers=admin_h,
-                      params={"mimeType": "text/plain"}).text
-    assert doc["content"] in text and text.startswith(doc["title"])
 
 
 def test_github_body_roundtrip(client, admin_h, ro_conn, org):
@@ -1215,22 +954,6 @@ def test_slack_users_info_resolves_author(client, admin_h, ro_conn):
     assert bad == {"ok": False, "error": "user_not_found"}
 
 
-def test_drive_in_owners_query(client, admin_h, ro_conn):
-    # real Drive supports `'<owner>' in owners`; the mock must filter by owner (email or name),
-    # not ignore the clause. (qst_0031's broken owner-lookup path.)
-    total = db_count(ro_conn, "google_drive")
-    owner = ro_conn.execute("SELECT author_email FROM gdrive_files LIMIT 1").fetchone()["author_email"]
-    expected = ro_conn.execute("SELECT count(*) FROM gdrive_files WHERE author_email=?", (owner,)).fetchone()[0]
-    j = client.get("/drive/v3/files", headers=admin_h,
-                   params={"q": f"'{owner}' in owners", "pageSize": 1000}).json()
-    n = len(j.get("files", []))
-    assert 0 < n < total and n == expected  # filtered to exactly this owner's files
-    # a non-owner returns nothing (clause honored, not ignored)
-    none = client.get("/drive/v3/files", headers=admin_h,
-                      params={"q": "'nobody-xyz@acme.com' in owners", "pageSize": 100}).json()
-    assert none.get("files", []) == []
-
-
 # --- Slack fidelity (#33) ---------------------------------------------------------------------
 #
 # Reported from building a filesystem-style Slack client against the mock. Slack answers an
@@ -1384,91 +1107,6 @@ def test_slack_search_all(client, admin_h):
     assert j["files"]["total"] == 0 and j["files"]["matches"] == []
 
 
-def test_google_batch_dispatches_subrequests(client, admin_h, ro_conn):
-    # google-api-python-client posts a multipart/mixed batch to /batch; the mock must dispatch each
-    # application/http sub-request in-process and return a multipart/mixed of sub-responses matched
-    # by Content-ID. Regression for the batch escaping to real Google (401). Build the batch body
-    # exactly like BatchHttpRequest does.
-    from email.generator import Generator
-    from email.mime.multipart import MIMEMultipart
-    from email.mime.nonmultipart import MIMENonMultipart
-    from email.parser import BytesParser
-    from io import StringIO
-
-    listed = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                        params={"maxResults": 2}).json().get("messages", [])
-    ids = [m["id"] for m in listed]
-    assert ids, "need at least one gmail message in the sample"
-
-    msg = MIMEMultipart("mixed")
-    setattr(msg, "_write_headers", lambda self: None)
-    for i, mid in enumerate(ids):
-        part = MIMENonMultipart("application", "http")
-        part["Content-Transfer-Encoding"] = "binary"
-        part["Content-ID"] = f"<base + {i}>"  # the format BatchHttpRequest uses
-        # format=full is the discriminator: a sub-request whose query is honored returns a payload;
-        # one whose query is dropped defaults to full too, so we assert the OPPOSITE below with
-        # format=minimal — see test_google_batch_honors_subrequest_query_params.
-        part.set_payload(f"GET /gmail/v1/users/me/messages/{mid}?format=full HTTP/1.1\r\n\r\n")
-        msg.attach(part)
-    fp = StringIO()
-    Generator(fp, mangle_from_=False).flatten(msg, unixfrom=False)
-    body, boundary = fp.getvalue(), msg.get_boundary()
-
-    r = client.post("/batch", headers={**admin_h, "Content-Type": f'multipart/mixed; boundary="{boundary}"'},
-                    content=body)
-    assert r.status_code == 200, r.text
-    assert "multipart/mixed" in r.headers["content-type"]
-    parsed = BytesParser().parsebytes(
-        b"Content-Type: " + r.headers["content-type"].encode() + b"\r\n\r\n" + r.content)
-    parts = parsed.get_payload()
-    assert len(parts) == len(ids)
-    for i, (mid, part) in enumerate(zip(ids, parts)):
-        assert part["Content-ID"] == f"<base + {i}>"          # echoed so the client can pair them
-        sub = part.get_payload(decode=False)
-        assert sub.startswith("HTTP/1.1 200")                  # dispatched with the admin token, not 401
-        assert mid in sub                                      # the message JSON came back
-
-
-def _batch_one(client, headers, mid, fmt, uri="/batch"):
-    """POST a one-message Gmail batch to `uri` (default /batch; /batch/gmail/v1 is the real Gmail
-    path) requesting `fmt`, and return the decoded sub-response JSON. Serialized exactly like
-    google-api-python-client's BatchHttpRequest."""
-    from email.generator import Generator
-    from email.mime.multipart import MIMEMultipart
-    from email.mime.nonmultipart import MIMENonMultipart
-    from email.parser import BytesParser
-    from io import StringIO
-
-    msg = MIMEMultipart("mixed")
-    setattr(msg, "_write_headers", lambda self: None)
-    part = MIMENonMultipart("application", "http")
-    part["Content-Transfer-Encoding"] = "binary"
-    part["Content-ID"] = "<b + 0>"
-    part.set_payload(f"GET /gmail/v1/users/me/messages/{mid}?format={fmt} HTTP/1.1\r\n\r\n")
-    msg.attach(part)
-    fp = StringIO()
-    Generator(fp, mangle_from_=False).flatten(msg, unixfrom=False)
-    r = client.post(uri, headers={**headers, "Content-Type": f'multipart/mixed; boundary="{msg.get_boundary()}"'},
-                    content=fp.getvalue())
-    assert r.status_code == 200, r.text
-    parsed = BytesParser().parsebytes(
-        b"Content-Type: " + r.headers["content-type"].encode() + b"\r\n\r\n" + r.content)
-    sub = parsed.get_payload()[0].get_payload(decode=False)
-    return json.loads(sub.split("\r\n\r\n", 1)[1])
-
-
-@pytest.mark.parametrize("uri", ["/batch", "/batch/gmail/v1"])
-def test_google_batch_honors_subrequest_query_params(client, admin_h, uri):
-    # The sub-request's query string must reach the dispatched handler. `format` is the tell: a
-    # dropped query defaults to full, so if the mock ignored it, `format=minimal` would still carry a
-    # payload. A batch-trusting client that caches these would cache bodyless messages otherwise.
-    mid = client.get("/gmail/v1/users/me/messages", headers=admin_h,
-                     params={"maxResults": 1}).json()["messages"][0]["id"]
-    assert "payload" in _batch_one(client, admin_h, mid, "full", uri)     # format=full honored
-    assert "payload" not in _batch_one(client, admin_h, mid, "minimal", uri)  # format=minimal honored
-
-
 def test_slack_replies_resolve_from_a_reply_ts(client, admin_h):
     # A search hit that lands on a REPLY yields that reply's ts; conversations.replies must return
     # the whole thread from it (Slack accepts any in-thread ts), not thread_not_found. The SAMPLE
@@ -1489,27 +1127,9 @@ def test_slack_replies_resolve_from_a_reply_ts(client, admin_h):
     assert "Rolled back" in texts               # the reply we searched for is in the same thread
 
 
-def test_user_cannot_fetch_others_private_gmail(client, tokens_yaml, admin_h, ro_conn):
-    # a private gmail doc owned by user B, fetched with user A's token -> 404
-    user_a, user_b = tokens_yaml["users"][0], tokens_yaml["users"][1]
-    doc = ro_conn.execute(
-        "SELECT doc_id FROM gmail_messages WHERE author_email=? LIMIT 1",
-        (user_b["email"],),
-    ).fetchone()
-    if doc is None:
-        pytest.skip("no gmail doc for user B in this subset")
-    from app import synth
-    hexid = synth.gmail_message_id(doc["doc_id"])   # served ids are hex, not dsids (#39)
-    ah = {"Authorization": f"Bearer {user_a['token']}"}
-    r = client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=ah)
-    # A may coincidentally be a recipient; assert admin can always read it
-    assert client.get(f"/gmail/v1/users/me/messages/{hexid}", headers=admin_h).status_code == 200
-    assert r.status_code in (200, 404)
-
-
 # --------------------------------------------------------------------------- Notion
 
-def _tok(tokens_yaml, email):
+def tok(tokens_yaml, email):
     return next(u["token"] for u in tokens_yaml["users"] if u["email"] == email)
 
 
@@ -1570,11 +1190,11 @@ def test_notion_unauth_is_401(client):
 def test_notion_acl_hides_group_doc_from_outsider(client, tokens_yaml):
     from app import synth
     pid = synth.notion_id("nt-secret")
-    outsider = _tok(tokens_yaml, "ava@acme.com")  # ava is engineering, not people
+    outsider = tok(tokens_yaml, "ava@acme.com")  # ava is engineering, not people
     r = client.get(f"/notion/v1/pages/{pid}", headers={"Authorization": f"Bearer {outsider}"})
     assert r.status_code == 404 and r.json()["code"] == "object_not_found"
     # the owner (hana, in people) can see it
-    owner = _tok(tokens_yaml, "hana@acme.com")
+    owner = tok(tokens_yaml, "hana@acme.com")
     assert client.get(f"/notion/v1/pages/{pid}",
                       headers={"Authorization": f"Bearer {owner}"}).status_code == 200
 
@@ -1905,166 +1525,6 @@ def test_s3_max_keys_zero_returns_empty_page_safely(big_bucket_client, big_bucke
     assert root.findtext(f"{{{S3NS}}}NextContinuationToken") is None
 
 
-# --- Google error envelope (#37) ------------------------------------------------------------
-#
-# Every case below was MEASURED against the live APIs with real OAuth credentials. The envelope is
-# per-family, not uniform:
-#
-#   family                       errors[]   status                 no Authorization header
-#   -----------------------------|----------|-----------------------|------------------------
-#   Drive v3                     | always   | auth failures only    | 403 PERMISSION_DENIED
-#   Gmail v1                     | always   | always                | 401 UNAUTHENTICATED
-#   Docs v1 / Sheets v4 / Slides | never    | always                | 401 UNAUTHENTICATED
-#
-# A bad bearer token is 401 UNAUTHENTICATED in every family.
-
-def _gerr(resp):
-    """The `error` object, or a clear failure naming what came back instead."""
-    body = resp.json()
-    assert "error" in body, f"expected a Google error envelope, got {body}"
-    return body["error"]
-
-
-def test_google_errors_use_googles_envelope(client, admin_h):
-    """`google-api-python-client` reads `error.message` to build HttpError, so `{"detail": …}` left
-    every error unreadable to the one client the mock exists to serve."""
-    r = client.get("/drive/v3/files", headers=admin_h, params={"fields": "totallyBogusField"})
-    assert r.status_code == 400
-    e = _gerr(r)
-    assert e["code"] == 400
-    assert e["message"] == "Invalid field selection totallyBogusField"
-    assert "detail" not in r.json()
-    # non-Google paths keep FastAPI's default envelope
-    assert "detail" in client.get("/no-such-route").json()
-
-
-def test_drive_errors_carry_the_legacy_errors_array(client, admin_h):
-    """Drive v3 always sends `errors[]` with a `reason` a client can branch on, and repeats the
-    message inside it. It does NOT send `status` for a parameter failure — measured."""
-    e = _gerr(client.get("/drive/v3/files", headers=admin_h, params={"fields": "nope"}))
-    assert e["errors"] == [{"message": "Invalid field selection nope", "domain": "global",
-                            "reason": "invalidParameter", "location": "fields",
-                            "locationType": "parameter"}]
-    assert "status" not in e, "Drive omits status on parameter failures"
-
-
-def test_editor_api_errors_carry_status_and_no_errors_array(client, admin_h):
-    """The editor APIs are the mirror image of Drive: `status`, never `errors[]` — measured."""
-    doc = _drive_find(client, admin_h, "Brand")["id"]
-    e = _gerr(client.get(f"/sheets/v4/spreadsheets/{doc}", headers=admin_h))
-    assert e["code"] == 404 and e["status"] == "NOT_FOUND"
-    assert e["message"] == "Requested entity was not found."
-    assert "errors" not in e
-
-
-def test_gmail_errors_carry_both(client, admin_h):
-    """Gmail sends `errors[]` AND `status` — measured, and the only family that does both."""
-    # a well-formed but unknown id; a non-hex one is 400 "Invalid id value" (see #39)
-    e = _gerr(client.get("/gmail/v1/users/me/messages/00000000deadbeef", headers=admin_h))
-    assert e["code"] == 404 and e["status"] == "NOT_FOUND"
-    assert e["message"] == "Requested entity was not found."
-    assert e["errors"][0]["reason"] == "notFound"
-
-
-# (path, params, code, status, reason, location) — one row per measured case.
-GOOGLE_ERROR_CASES = [
-    ("/drive/v3/files", {"fields": "bogus"}, 400, None, "invalidParameter", "fields"),
-    ("/drive/v3/files", {"orderBy": "bogusKey"}, 400, None, "invalid", "orderBy"),
-    ("/drive/v3/files/no-such-file", {}, 404, None, "notFound", "fileId"),
-    ("/drive/v3/about", {}, 400, None, "required", "fields"),
-    ("/drive/v3/about", {"fields": "storageQuoat"}, 400, None, "invalidParameter", "fields"),
-    ("/gmail/v1/users/me/messages/00000000deadbeef", {}, 404, "NOT_FOUND", "notFound", None),
-    ("/gmail/v1/users/me/labels/NO_SUCH", {}, 404, "NOT_FOUND", "notFound", None),
-]
-
-
-@pytest.mark.parametrize("path, params, code, status, reason, location", GOOGLE_ERROR_CASES)
-def test_google_error_reasons_match_the_real_api(client, admin_h, path, params, code, status,
-                                                 reason, location):
-    r = client.get(path, headers=admin_h, params=params)
-    assert r.status_code == code
-    e = _gerr(r)
-    assert e["code"] == code
-    assert e.get("status") == status
-    err0 = e["errors"][0]
-    assert err0["reason"] == reason
-    assert err0["domain"] == "global"
-    assert err0.get("location") == location
-    if location is not None:
-        assert err0["locationType"] == "parameter"
-
-
-def test_drive_not_found_names_the_file_id(client, admin_h):
-    """Measured: `File not found: {id}.` — the id is in the message, so a batch caller can tell
-    which of its requests failed."""
-    e = _gerr(client.get("/drive/v3/files/abc123xyz", headers=admin_h))
-    assert e["message"] == "File not found: abc123xyz."
-
-
-def test_drive_export_requires_mime_type_with_googles_wording(client, admin_h):
-    doc = _drive_find(client, admin_h, "Brand")["id"]
-    e = _gerr(client.get(f"/drive/v3/files/{doc}/export", headers=admin_h))
-    assert e["code"] == 400 and e["message"] == "Required parameter: mimeType"
-    assert e["errors"][0] == {"message": "Required parameter: mimeType", "domain": "global",
-                              "reason": "required", "location": "mimeType",
-                              "locationType": "parameter"}
-
-
-@pytest.mark.parametrize("path, reason, location", [
-    ("/drive/v3/files/{pdf}/export?mimeType=text/plain", "fileNotExportable", None),
-    ("/drive/v3/files/{doc}?alt=media", "fileNotDownloadable", "alt"),
-])
-def test_drive_403s_carry_their_own_reasons(client, admin_h, path, reason, location):
-    doc = _drive_find(client, admin_h, "Brand")["id"]
-    pdf = _drive_find(client, admin_h, "Whitepaper")["id"]
-    e = _gerr(client.get(path.format(doc=doc, pdf=pdf), headers=admin_h))
-    assert e["code"] == 403
-    assert e["errors"][0]["reason"] == reason
-    assert e["errors"][0].get("location") == location
-
-
-BAD_TOKEN = {"Authorization": "Bearer not-a-real-token"}
-
-
-@pytest.mark.parametrize("path", ["/drive/v3/files", "/gmail/v1/users/me/profile",
-                                  "/sheets/v4/spreadsheets/x", "/docs/v1/documents/x",
-                                  "/slides/v1/presentations/x"])
-def test_a_bad_token_is_unauthenticated_everywhere(client, path):
-    """Measured: every family answers a present-but-invalid bearer with 401 UNAUTHENTICATED, and
-    the short "Invalid Credentials" lives in `errors[0]` while the top message is the long form."""
-    r = client.get(path, headers=BAD_TOKEN)
-    assert r.status_code == 401
-    e = _gerr(r)
-    assert e["code"] == 401 and e["status"] == "UNAUTHENTICATED"
-    assert e["message"].startswith("Request had invalid authentication credentials.")
-    if "errors" in e:
-        assert e["errors"][0]["message"] == "Invalid Credentials"
-        assert e["errors"][0]["reason"] == "authError"
-        assert e["errors"][0]["location"] == "Authorization"
-        assert e["errors"][0]["locationType"] == "header"
-
-
-@pytest.mark.parametrize("path, code, status", [
-    ("/drive/v3/files", 403, "PERMISSION_DENIED"),        # Drive accepts API keys, so anonymous
-    ("/sheets/v4/spreadsheets/x", 403, "PERMISSION_DENIED"),  # ...is an "unregistered caller"
-    ("/gmail/v1/users/me/profile", 401, "UNAUTHENTICATED"),   # OAuth-only APIs say the
-    ("/docs/v1/documents/x", 401, "UNAUTHENTICATED"),         # ...credentials are missing
-    ("/slides/v1/presentations/x", 401, "UNAUTHENTICATED"),
-])
-def test_a_missing_header_differs_by_family(client, path, code, status):
-    """The surprise, measured: no `Authorization` header at all is NOT uniformly 401. Drive and
-    Sheets answer 403 PERMISSION_DENIED, Gmail and the Docs/Slides APIs answer 401. A bad token is
-    401 everywhere — so the two cases are genuinely distinct and the mock conflated them."""
-    r = client.get(path)
-    assert r.status_code == code
-    e = _gerr(r)
-    assert e["code"] == code and e["status"] == status
-    if code == 403:
-        assert "unregistered callers" in e["message"]
-    else:
-        assert "missing required authentication credential" in e["message"]
-
-
 def test_atlassian_errors_use_atlassian_envelope(client):
     # atlassian-python-api's Confluence client does response.json()["message"] on any error, so the
     # mock must shape /atlassian errors like Atlassian Cloud (message + statusCode), not {"detail"}.
@@ -2170,437 +1630,6 @@ def test_slack_api_test_has_typed_response_schema(client):
     schema = op["responses"]["200"]["content"]["application/json"]["schema"]
     assert schema != {}
     assert "$ref" in schema or schema.get("type") in ("object", "array")
-
-
-# --- Gmail: typed response schema, unchanged responses ------------------------------------
-
-def test_gmail_messages_has_typed_response_schema(client):
-    op = client.get("/openapi.json").json()["paths"]["/gmail/v1/users/{user_id}/messages"]["get"]
-    schema = op["responses"]["200"]["content"]["application/json"]["schema"]
-    assert schema != {}
-
-
-def test_gmail_responses_unchanged_by_enrichment(client, admin_h):
-    lst = client.get("/gmail/v1/users/me/messages", headers=admin_h).json()
-    assert "messages" in lst and "resultSizeEstimate" in lst
-    if lst["messages"]:
-        mid = lst["messages"][0]["id"]
-        msg = client.get(f"/gmail/v1/users/me/messages/{mid}", params={"format": "full"},
-                         headers=admin_h).json()
-        for k in ("id", "threadId", "labelIds", "snippet", "internalDate", "sizeEstimate", "payload"):
-            assert k in msg, f"gmail message missing {k} (fidelity regression)"
-
-
-# --- OpenAPI enrichment: drive ------------------------------------------------------------
-
-def test_drive_files_has_typed_response_schema(client):
-    op = client.get("/openapi.json").json()["paths"]["/drive/v3/files"]["get"]
-    schema = op["responses"]["200"]["content"]["application/json"]["schema"]
-    assert schema != {}
-
-
-def _drive_find(client, admin_h, name_substr):
-    j = client.get("/drive/v3/files", params={"q": f"name contains '{name_substr}'"},
-                   headers=admin_h).json()
-    return j["files"][0] if j.get("files") else None
-
-
-def test_drive_responses_unchanged_by_enrichment(client, admin_h):
-    lst = client.get("/drive/v3/files", headers=admin_h).json()
-    assert lst["kind"] == "drive#fileList" and "files" in lst
-    doc = _drive_find(client, admin_h, "Brand")
-    assert doc is not None
-    full = client.get(f"/drive/v3/files/{doc['id']}", headers=admin_h).json()
-    for k in ("kind", "id", "name", "mimeType", "createdTime", "modifiedTime", "owners",
-              "webViewLink", "capabilities"):
-        assert k in full, f"drive file missing {k} (fidelity regression)"
-
-
-def test_drive_export_and_media_stay_non_json(client, admin_h):
-    # A native doc exports as PlainTextResponse; response_model must NOT be attached to these.
-    doc = _drive_find(client, admin_h, "Brand")
-    exp = client.get(f"/drive/v3/files/{doc['id']}/export",
-                     params={"mimeType": "text/plain"}, headers=admin_h)
-    assert exp.status_code == 200 and "application/json" not in exp.headers["content-type"]
-    # A binary (pdf) downloads raw via alt=media.
-    pdf = _drive_find(client, admin_h, "Whitepaper")
-    med = client.get(f"/drive/v3/files/{pdf['id']}", params={"alt": "media"}, headers=admin_h)
-    assert med.status_code == 200 and "application/json" not in med.headers["content-type"]
-
-
-# --- Drive fidelity: measured divergences from real Google Drive (issue #23) ---------------
-#
-# Each case below was diffed against https://www.googleapis.com/drive/v3 with equivalent
-# credentials; the mock's old behaviour returned 200 with wrong/unfiltered data, so a consumer
-# could not tell anything was off.
-
-FOLDER_MIME = "application/vnd.google-apps.folder"
-DOC_MIME = "application/vnd.google-apps.document"
-
-
-def _drive_ids(client, headers, **params):
-    j = client.get("/drive/v3/files", headers=headers, params=params).json()
-    return [f["id"] for f in j.get("files", [])]
-
-
-def test_drive_shared_with_me_partitions_by_owner(client, tokens_yaml):
-    """`q=sharedWithMe=true` must return only items shared with the caller by someone else, and
-    `false` must exclude them — real Drive's "Shared with me" is the only way to enumerate those.
-    The mock used to ignore the clause, so both returned the caller's whole visible corpus."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    all_ids = set(_drive_ids(client, mia, q="trashed=false", pageSize=100))
-    shared = set(_drive_ids(client, mia, q="sharedWithMe=true and trashed=false", pageSize=100))
-    own = set(_drive_ids(client, mia, q="sharedWithMe=false and trashed=false", pageSize=100))
-    assert shared and own                      # SAMPLE gives mia both her own and others' files
-    assert shared != own and not (shared & own)
-    assert shared | own == all_ids             # together they partition the visible corpus
-    # mia authored "Brand guidelines v3"; it is hers, not shared with her
-    brand = _drive_find(client, mia, "Brand")["id"]
-    assert brand in own and brand not in shared
-
-
-def test_drive_shared_items_carry_shared_with_me_time(client, tokens_yaml):
-    """Real Drive populates `sharedWithMeTime` only on items shared with the caller, and omits
-    `parents` on them — so its presence is how a client classifies one. Filtering on
-    `sharedWithMe` while never emitting the field left a row that the filter calls shared unable to
-    say so itself."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    shared = client.get("/drive/v3/files", headers=mia,
-                        params={"q": "sharedWithMe=true and trashed=false",
-                                "pageSize": 100}).json()["files"]
-    own = client.get("/drive/v3/files", headers=mia,
-                     params={"q": "sharedWithMe=false and trashed=false",
-                             "pageSize": 100}).json()["files"]
-    assert shared and own
-    assert all(f["sharedWithMeTime"] for f in shared), "every shared item needs the timestamp"
-    assert all("sharedWithMeTime" not in f for f in own), "an item you own was never shared with you"
-    # folders come out of the same filter, so they must answer the same way
-    assert any(f["mimeType"] == FOLDER_MIME for f in shared)
-    # and files.get agrees with the listing
-    one = shared[0]
-    assert client.get(f"/drive/v3/files/{one['id']}", headers=mia).json() == one
-
-
-def test_drive_shared_with_me_time_needs_a_caller(client, admin_h):
-    """The admin/service token is not a Drive user, so nothing was shared *with* it — no timestamp
-    to invent. `orderBy` on the field still answers (all-equal keys), as real Drive does for nulls."""
-    files = client.get("/drive/v3/files", headers=admin_h, params={"pageSize": 20}).json()["files"]
-    assert files and all("sharedWithMeTime" not in f for f in files)
-    assert client.get("/drive/v3/files", headers=admin_h,
-                      params={"pageSize": 5, "orderBy": "sharedWithMeTime"}).status_code == 200
-
-
-def test_drive_order_by_shared_with_me_time(client, tokens_yaml):
-    """The mock models the relation this key sorts on (owner vs caller), so it sorts rather than
-    400s — unlike the view/modify-by-me timestamps, which have no counterpart here at all."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    r = client.get("/drive/v3/files", headers=mia,
-                   params={"q": "sharedWithMe=true", "pageSize": 100,
-                           "orderBy": "sharedWithMeTime desc"})
-    assert r.status_code == 200
-    times = [f["sharedWithMeTime"] for f in r.json()["files"]]
-    assert times == sorted(times, reverse=True)
-
-
-def test_drive_owned_by_me_reflects_the_caller(client, tokens_yaml):
-    """`ownedByMe` is per-caller in real Drive; the mock reported False for every file."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    assert _drive_find(client, mia, "Brand")["ownedByMe"] is True
-    assert _drive_find(client, mia, "Whitepaper")["ownedByMe"] is False
-
-
-def test_drive_order_by_sorts_the_result(client, admin_h):
-    """`orderBy` was accepted and never applied — silent, so a client that relies on server-side
-    ordering appears to work against the mock and misbehaves against production."""
-    names = [f["name"] for f in client.get(
-        "/drive/v3/files", headers=admin_h,
-        params={"q": "trashed=false", "pageSize": 100, "orderBy": "name",
-                "fields": "files(name)"}).json()["files"]]
-    # Drive collates names case-insensitively (folder names in the SAMPLE are lowercase, file
-    # names are not, so a case-sensitive sort would put every folder last)
-    assert names == sorted(names, key=str.casefold)
-    desc = [f["name"] for f in client.get(
-        "/drive/v3/files", headers=admin_h,
-        params={"q": "trashed=false", "pageSize": 100, "orderBy": "name desc",
-                "fields": "files(name)"}).json()["files"]]
-    assert desc == sorted(names, key=str.casefold, reverse=True)
-    mods = [f["modifiedTime"] for f in client.get(
-        "/drive/v3/files", headers=admin_h,
-        params={"q": "trashed=false", "pageSize": 100, "orderBy": "modifiedTime desc",
-                "fields": "files(modifiedTime)"}).json()["files"]]
-    assert mods == sorted(mods, reverse=True)
-
-
-def test_drive_order_by_paginates_in_sorted_order(client, admin_h):
-    """A sort must span the whole result set, not sort each page in isolation."""
-    everything = [f["name"] for f in client.get(
-        "/drive/v3/files", headers=admin_h,
-        params={"pageSize": 100, "orderBy": "name", "fields": "files(name)"}).json()["files"]]
-    paged, token = [], None
-    while True:
-        p = {"pageSize": 2, "orderBy": "name", "fields": "files(name),nextPageToken"}
-        if token:
-            p["pageToken"] = token
-        j = client.get("/drive/v3/files", headers=admin_h, params=p).json()
-        paged += [f["name"] for f in j["files"]]
-        token = j.get("nextPageToken")
-        if not token:
-            break
-    assert paged == everything == sorted(everything, key=str.casefold)
-
-
-def test_drive_order_by_does_not_change_the_rows_themselves(client, admin_h):
-    """Sorting builds the whole result set to order it, and defers the per-page `shared` lookup —
-    so the served objects must still be identical to the unsorted ones, field for field."""
-    plain = {f["id"]: f for f in client.get(
-        "/drive/v3/files", headers=admin_h, params={"pageSize": 100}).json()["files"]}
-    sorted_ = {f["id"]: f for f in client.get(
-        "/drive/v3/files", headers=admin_h,
-        params={"pageSize": 100, "orderBy": "modifiedTime desc"}).json()["files"]}
-    assert plain and plain == sorted_
-    assert any(f["shared"] for f in plain.values())   # ...and `shared` is really resolved
-
-
-def test_drive_order_by_rejects_keys_it_cannot_honor(client, admin_h):
-    """Real Drive 400s an undocumented sort key. The mock models no per-caller view/share
-    timestamps, so those documented keys are rejected loudly rather than silently ignored."""
-    for bad in ("bogusKey", "name descending", "viewedByMeTime"):
-        r = client.get("/drive/v3/files", headers=admin_h, params={"orderBy": bad})
-        assert r.status_code == 400, f"orderBy={bad!r} should 400, got {r.status_code}"
-    ok = client.get("/drive/v3/files", headers=admin_h,
-                    params={"orderBy": "folder,name desc", "pageSize": 5})
-    assert ok.status_code == 200
-
-
-def test_drive_invalid_fields_mask_is_rejected(client, admin_h):
-    """An unknown field name used to be accepted and yield empty file objects (200 {}), so a typo
-    or a stale field name in a consumer's mask passed every mock-backed test and 400d in
-    production."""
-    r = client.get("/drive/v3/files", headers=admin_h,
-                   params={"pageSize": 1, "fields": "files(totallyBogusField)"})
-    assert r.status_code == 400
-    assert "totallyBogusField" in r.json()["error"]["message"]
-    bad_top = client.get("/drive/v3/files", headers=admin_h,
-                         params={"pageSize": 1, "fields": "bogusTop,files(id)"})
-    assert bad_top.status_code == 400
-    # a documented field the mock does not synthesize is still valid (real Drive omits it, 200)
-    ok = client.get("/drive/v3/files", headers=admin_h,
-                    params={"pageSize": 1, "fields": "files(id,thumbnailLink,capabilities/canEdit)"})
-    assert ok.status_code == 200 and "thumbnailLink" not in ok.json()["files"][0]
-
-
-def test_drive_get_honors_the_fields_mask(client, admin_h):
-    """The same projection requested two ways must give the same object; files.get ignored the
-    mask entirely and added keys nobody asked for."""
-    mask = "id,name,mimeType,size,modifiedTime,webViewLink"
-    row = client.get("/drive/v3/files", headers=admin_h,
-                     params={"q": "name contains 'Brand'", "pageSize": 1,
-                             "fields": f"files({mask})"}).json()["files"][0]
-    got = client.get(f"/drive/v3/files/{row['id']}", headers=admin_h,
-                     params={"fields": mask}).json()
-    assert got == row
-    r = client.get(f"/drive/v3/files/{row['id']}", headers=admin_h,
-                   params={"fields": "totallyBogusField"})
-    assert r.status_code == 400
-
-
-def test_drive_folders_are_found_by_mime_type(client, admin_h):
-    """Folders were returned by `'root' in parents` but invisible to `mimeType='…folder'`, so a
-    crawler indexing folders by type concluded the account had none."""
-    by_parent = _drive_ids(client, admin_h, q="'root' in parents", pageSize=100)
-    by_mime = _drive_ids(client, admin_h, q=f"mimeType='{FOLDER_MIME}'", pageSize=100)
-    assert by_parent and set(by_mime) == set(by_parent)
-    # and the negation excludes them
-    not_folders = _drive_ids(client, admin_h, q=f"mimeType!='{FOLDER_MIME}'", pageSize=100)
-    assert not set(not_folders) & set(by_parent)
-
-
-def test_drive_folders_honor_the_fields_projection(client, admin_h):
-    """Synthesized folder rows bypassed the projection: `files(id,name)` returned 18 keys."""
-    for q in ("'root' in parents", f"mimeType='{FOLDER_MIME}'"):
-        files = client.get("/drive/v3/files", headers=admin_h,
-                           params={"q": q, "pageSize": 5, "fields": "files(id,name)"}).json()["files"]
-        assert files and all(set(f) == {"id", "name"} for f in files), q
-
-
-def test_drive_folders_match_the_same_q_clauses_as_files(client, admin_h):
-    """Folders now flow through `_drive_q_match`, so every clause that should match one does."""
-    folders = client.get("/drive/v3/files", headers=admin_h,
-                         params={"q": "'root' in parents", "pageSize": 100,
-                                 "fields": "files(id,name)"}).json()["files"]
-    one = folders[0]
-    hit = _drive_ids(client, admin_h, q=f"name contains '{one['name']}' and mimeType='{FOLDER_MIME}'")
-    assert one["id"] in hit
-    # a folder is not trashed, so trashed=true excludes it
-    assert one["id"] not in _drive_ids(client, admin_h, q=f"mimeType='{FOLDER_MIME}' and trashed=true")
-
-
-def test_drive_folder_permissions_resolve(client, admin_h):
-    """A folder id is a first-class file id in real Drive: files.get and permissions.list both
-    answer for it. permissions.list 404d because folders are not stored as rows."""
-    folder = client.get("/drive/v3/files", headers=admin_h,
-                        params={"q": "'root' in parents", "pageSize": 1}).json()["files"][0]
-    got = client.get(f"/drive/v3/files/{folder['id']}", headers=admin_h)
-    assert got.status_code == 200 and got.json()["mimeType"] == FOLDER_MIME
-    perms = client.get(f"/drive/v3/files/{folder['id']}/permissions", headers=admin_h)
-    assert perms.status_code == 200 and perms.json()["permissions"]
-
-
-def test_drive_native_docs_report_size(client, admin_h):
-    """Google populates `size` for binary content *and for Docs Editors files*; the mock omitted
-    it on native rows, which taught implementors something false about the API."""
-    doc = _drive_find(client, admin_h, "Brand")
-    assert doc["mimeType"] == DOC_MIME
-    assert int(doc["size"]) > 0
-    assert "md5Checksum" not in doc          # real Drive omits checksums on native files
-    folder = client.get("/drive/v3/files", headers=admin_h,
-                        params={"q": "'root' in parents", "pageSize": 1}).json()["files"][0]
-    assert "size" not in folder              # ...but not for folders or shortcuts
-
-
-# --- Drive about.get -----------------------------------------------------------------------
-#
-# `about` answers "who am I and how much space do I use" — the call a Drive client makes first,
-# and the one the mock had no route for at all (404). Its contract is unusual: `fields` is
-# mandatory, and the response carries only what the mask asked for.
-
-ABOUT = "/drive/v3/about"
-SHEET_MIME = "application/vnd.google-apps.spreadsheet"
-
-
-def _about(client, headers, fields):
-    return client.get(ABOUT, headers=headers, params={"fields": fields})
-
-
-def test_drive_about_requires_a_fields_mask(client, admin_h):
-    """Real Drive 400s `about.get` with no `fields` — this resource has no default projection.
-    Serving a full body instead would let a client ship a call that fails in production."""
-    r = client.get(ABOUT, headers=admin_h)
-    assert r.status_code == 400
-    assert "fields" in r.json()["error"]["message"]
-
-
-def test_drive_about_rejects_an_unknown_field(client, admin_h):
-    """Same rule as the `files` masks: a typo 400s rather than quietly matching nothing."""
-    assert _about(client, admin_h, "storageQuoat").status_code == 400
-    assert _about(client, admin_h, "storageQuota").status_code == 200
-
-
-def test_drive_about_rejects_a_mask_that_selects_nothing(client, admin_h):
-    """`fields=,` clears the required-mask check but names no field. Falling through to "no
-    projection" would answer a request for nothing with the entire resource."""
-    r = client.get(ABOUT, headers=admin_h, params={"fields": ","})
-    assert r.status_code == 400
-
-
-def test_drive_about_needs_auth(client):
-    # no header at all -> 403 on Drive (an unregistered caller); a bad token -> 401
-    assert client.get(ABOUT, params={"fields": "user"}).status_code == 403
-    bad = {"Authorization": "Bearer nope"}
-    assert client.get(ABOUT, params={"fields": "user"}, headers=bad).status_code == 401
-    # auth is resolved before the mask, as real Drive does — a missing mask on a bad token is 401
-    assert client.get(ABOUT, headers=bad).status_code == 401
-
-
-def test_drive_about_serves_only_the_requested_fields(client, admin_h):
-    """Unlike `files.list` — whose typed response model always carries `kind` — `about` projects
-    strictly, which is what real Drive does: ask for `user` and `user` is all you get."""
-    j = _about(client, admin_h, "user").json()
-    assert set(j) == {"user"}
-    assert set(_about(client, admin_h, "user,storageQuota").json()) == {"user", "storageQuota"}
-
-
-def test_drive_about_nested_mask_selects_its_parent(client, admin_h):
-    """`storageQuota/limit` selects `storageQuota`, the same rule every other mask in this mock
-    follows — one projection depth, applied consistently."""
-    j = _about(client, admin_h, "storageQuota/limit").json()
-    assert set(j) == {"storageQuota"}
-    assert "usage" in j["storageQuota"]
-
-
-def test_drive_about_user_is_the_caller(client, tokens_yaml):
-    """`about.user` is the authenticated user, so `me` is true — the opposite of the same object
-    read as a file's `owners` entry, where it describes someone else."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    u = _about(client, mia, "user").json()["user"]
-    assert u["kind"] == "drive#user"
-    assert u["emailAddress"] == "mia@acme.com"
-    assert u["me"] is True
-    # the file resource keeps its own answer: mia as an owner is not "me" to the object itself
-    assert _drive_find(client, mia, "Brand")["owners"][0]["me"] is False
-
-
-def test_drive_about_admin_token_reports_a_concrete_address(client, admin_h):
-    """The admin/service token is not a Drive person; real Drive still never reports a placeholder
-    here, so the service identity stands in — as `gmail.users.getProfile` already does."""
-    u = _about(client, admin_h, "user").json()["user"]
-    assert "@" in u["emailAddress"] and u["me"] is True
-
-
-def test_drive_about_usage_matches_the_sizes_files_list_serves(client, tokens_yaml):
-    """storageQuota and files.list are two views of one corpus. If they disagree, a client cannot
-    reconcile "how much space do I use" with "what is in my Drive"."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    quota = _about(client, mia, "storageQuota").json()["storageQuota"]
-    files = client.get("/drive/v3/files", headers=mia,
-                       params={"pageSize": 100, "fields": "files(size)"}).json()["files"]
-    listed = sum(int(f["size"]) for f in files if "size" in f)   # folders carry no size
-    assert listed > 0
-    assert int(quota["usageInDrive"]) == listed
-    assert quota["usage"] == quota["usageInDrive"]   # the mock stores nothing outside Drive
-    assert int(quota["limit"]) == 2199023255552     # 2 TiB
-    assert int(quota["usageInDriveTrash"]) == 0     # SAMPLE trashes nothing
-
-
-def test_drive_about_usage_is_scoped_to_the_caller(client, admin_h, tokens_yaml):
-    """A scoped token must not be told the weight of a corpus it cannot read."""
-    mia = {"Authorization": f"Bearer {_tok(tokens_yaml, 'mia@acme.com')}"}
-    mine = int(_about(client, mia, "storageQuota").json()["storageQuota"]["usage"])
-    everything = int(_about(client, admin_h, "storageQuota").json()["storageQuota"]["usage"])
-    assert 0 < mine < everything
-
-
-def test_drive_about_export_formats_are_honoured_by_files_export(client, admin_h):
-    """Advertising a target that `files.export` refuses would be worse than advertising nothing:
-    a client reads this map to decide what to ask for."""
-    formats = _about(client, admin_h, "exportFormats").json()["exportFormats"]
-    doc = _drive_find(client, admin_h, "Brand")
-    assert doc["mimeType"] == DOC_MIME and formats[DOC_MIME]
-    for target in formats[DOC_MIME]:
-        r = client.get(f"/drive/v3/files/{doc['id']}/export", headers=admin_h,
-                       params={"mimeType": target})
-        assert r.status_code == 200, target
-    # every native type the mock serves is covered; the folder type is not exportable anywhere
-    assert set(formats) == {DOC_MIME, SHEET_MIME,
-                            "application/vnd.google-apps.presentation"}
-    assert "text/csv" in formats[SHEET_MIME]
-
-
-def test_drive_about_shared_drive_fields_agree_with_the_drives_listing(client, admin_h):
-    """The mock's corpus is all My Drive and `/drive/v3/drives` is empty, so every shared-drive
-    field has to say the same thing rather than hinting at a capability that isn't there."""
-    j = _about(client, admin_h, "*").json()
-    assert client.get("/drive/v3/drives", headers=admin_h).json()["drives"] == []
-    assert j["canCreateDrives"] is False and j["canCreateTeamDrives"] is False
-    assert j["driveThemes"] == [] and j["teamDriveThemes"] == []
-
-
-def test_drive_about_star_serves_the_whole_resource(client, admin_h):
-    j = _about(client, admin_h, "*").json()
-    assert j["kind"] == "drive#about"
-    assert j["appInstalled"] is False
-    assert {"user", "storageQuota", "importFormats", "exportFormats", "maxImportSizes",
-            "maxUploadSize", "folderColorPalette"} <= set(j)
-    # folderColorRgb is a documented file field, so the palette a client picks from must be real
-    assert all(re.fullmatch(r"#[0-9a-f]{6}", c) for c in j["folderColorPalette"])
-    assert DOC_MIME in j["importFormats"]["text/plain"]
-
-
-def test_drive_about_appears_in_the_openapi_spec(client):
-    """The OpenAPI→MCP bridge builds its tools from the spec, so a route the spec omits is a route
-    no generated client can reach."""
-    op = client.get("/openapi.json").json()["paths"][ABOUT]["get"]
-    assert {p["name"] for p in op["parameters"]} == {"fields"}
 
 
 # --- Notion: typed response schema ---------------------------------------------------------
